+++ Sources/Machine/py_mps_periodic.hpp
@@ -30,15 +30,10 @@ namespace py = pybind11;
 namespace netket {
 
 void AddMpsPeriodic(py::module &subm) {
-  py::class_<MPSPeriodic<false>, AbstractMachine>(subm, "MPSPeriodic")
-      .def(py::init<const AbstractHilbert &, double, int>(),
+  py::class_<MPSPeriodic, AbstractMachine>(subm, "MPSPeriodic")
+      .def(py::init<const AbstractHilbert &, int, bool, int>(),
            py::keep_alive<1, 2>(), py::arg("hilbert"), py::arg("bond_dim"),
-           py::arg("symperiod") = -1);
-
-  py::class_<MPSPeriodic<true>, AbstractMachine>(subm, "MPSPeriodicDiagonal")
-      .def(py::init<const AbstractHilbert &, double, int>(),
-           py::keep_alive<1, 2>(), py::arg("hilbert"), py::arg("bond_dim"),
-           py::arg("symperiod") = -1);
+           py::arg("diag") = false, py::arg("symperiod") = -1);
 }
 
 }  // namespace netket
